+++ packages/lib/src/helpers/read.ts
@@ -1,9 +1,13 @@
-import import_ from '@brillout/import';
-import { IPlugin } from '@connery-io/sdk';
+import { AbstractPluginFactory, PluginFactoryContext } from '@connery-io/sdk';
 
-export async function readPluginDefinitionFile(pluginDefinitionPath: string): Promise<IPlugin> {
-  const importedModule = await import_(pluginDefinitionPath);
-  const PluginClass = importedModule.default;
-  const pluginInstance: IPlugin = new PluginClass();
-  return pluginInstance;
+export async function getPluginFactory(
+  pluginDefinitionPath: string,
+  pluginFactoryContext: PluginFactoryContext,
+): Promise<AbstractPluginFactory> {
+  const importedModule = await import(pluginDefinitionPath);
+  const pluginFactoryClass = importedModule.default.default as {
+    new (context: PluginFactoryContext): AbstractPluginFactory;
+  };
+  const pluginFactoryInstance = new pluginFactoryClass(pluginFactoryContext);
+  return pluginFactoryInstance;
 }